Reorganize cards-resource structure
