+++ src/lib/helpers/htmlContent.ts
