+++ src/lib/helpers/txtIterator.ts
