+++ src/routes/(cards-resources)/+layout.svelte
@@ -1,7 +1,5 @@
 <script>
-
 	import Linkbar from "$lib/component/Linkbar.svelte";
-
 </script>
 <div class="lg:w-[1024px] md:max-w-5xl mb-12">
     <h1 class="text-textprimary font-bold tracking-wide text-4xl mb-4">Hare (ハレ)</h1>
